Implement correct bytewise comparison
Handle cases where the prefix determines whether or not the rest
of the byte string fits in the range. For example:

Low = Included([0, 0, 0])
High = Excluded([0, 1, 0])
Key = [0, 0, 1]

Applying comparison bytewise causes the last byte of the key to
compare as greater than the last byte of the high bound, which is wrong.