Reduce range iteration to linear time in key length
We actually only need to process the `min` and `max` readers
exactly once on the boundary of the traversal (sides of the triangle).
This means we can just mutate them in place, and compare only
`byte::Array` slices instead of the entire min/max/key!

This also removes one more required connection between the `Read`
and `Write` traits, *and* allows us to support efficient value-only
range iteration, so we can finally combine it with the prefix
iterator.