Remove length field from fixed writer
Mostly for linearizable range scans, which take a buffer of
`Vec<(K::Write, u64)>`. We don't want to use `Vec<(K, u64)>` because
the optimistic validation compares keys using `Ord`, which may
not agree with the lexicographic ordering used by the iterators.

This does mean that the length is double-counted for dynamic writers,
which do track their length.